requirements.txt
attrs==25.3.0
aws-cdk.assets==1.204.0
aws-cdk.aws-applicationautoscaling==1.204.0
aws-cdk.aws-autoscaling-common==1.204.0
aws-cdk.aws-cloudwatch==1.204.0
aws-cdk.aws-codeguruprofiler==1.204.0
aws-cdk.aws-codestarnotifications==1.204.0
aws-cdk.aws-ec2==1.204.0
aws-cdk.aws-ecr==1.204.0
aws-cdk.aws-ecr-assets==1.204.0
aws-cdk.aws-efs==1.204.0
aws-cdk.aws-events==1.204.0
aws-cdk.aws-iam==1.204.0
aws-cdk.aws-kms==1.204.0
aws-cdk.aws-lambda==1.204.0
aws-cdk.aws-logs==1.204.0
aws-cdk.aws-s3==1.204.0
aws-cdk.aws-s3-assets==1.204.0
aws-cdk.aws-signer==1.204.0
aws-cdk.aws-sns==1.204.0
aws-cdk.aws-sqs==1.204.0
aws-cdk.aws-ssm==1.204.0
aws-cdk.cloud-assembly-schema==1.204.0
aws-cdk.core==1.204.0
aws-cdk.cx-api==1.204.0
aws-cdk.region-info==1.204.0
aws-xray-sdk
cattrs==24.1.3
constructs==3.4.344
importlib_resources==6.5.2
jsii==1.111.0
publication==0.0.3
python-dateutil==2.9.0.post0
six==1.17.0
typeguard==2.13.3
typing_extensions==4.13.1
